Implemented algorithm methods
