+++ TSPEADemo/go-backend/src/main/algorithm-runner.go
@@ -1,0 +1,40 @@
+package main
+
+import (
+	"fmt"
+	"time"
+	"math/rand"
+)
+
+type AlgorithmRunner struct {
+	Running bool
+	Value   float32
+}
+
+// TODO implement multi threading here
+
+// Start using a pointer so we can run a go routine on the actual travelr
+func startAlgorithmRunner() *AlgorithmRunner {
+
+	// Create new traveller and set to running
+	// TODO Should accept the config from the client side
+
+	traveler := AlgorithmRunner{Running: true}
+	fmt.Println("start")
+	go func() {
+		// TODO actually do the genetic travelling here
+		for traveler.Running == true {
+			traveler.Value = rand.Float32()
+			fmt.Println(traveler.Value)
+
+			// Go is to fast so we sleep :)
+			time.Sleep(1000*time.Millisecond)
+		}
+	}()
+	return &traveler
+
+}
+
+func (traveler *AlgorithmRunner) stop() {
+	traveler.Running = false
+}
